updated with more changes
