+++ _docker/src/general/backend/config.yaml
@@ -5,8 +5,12 @@ server:
     - levels: "info|error|debug"
   sources:
     - path: "../frontend/tests/playwright-files"
+      config:
+        defaultEnabled: true
     - path: "."
       name: "docker"
+      config:
+        defaultEnabled: true
 frontend:
   name: "Graham's Filebrowser"
   disableDefaultLinks: true
