+++ frontend/src/components/ContextMenu.vue
@@ -140,7 +140,7 @@ export default {
   },
   computed: {
     isMobileDevice() {
-      return /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent);
+      return /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent) || getters.isMobile();
     },
     noItems() {
       return !this.showEdit && !this.showSave && !this.showDelete;
