+++ frontend/tests-playwright/noauth/share.spec.ts
@@ -65,37 +65,3 @@ test("share download single file", async ({ page, checkForErrors, context }) => 
   await expect(popup).toHaveText("Downloading...");
   checkForErrors();
 });
-
-test("share download multiple files", async ({ page, checkForErrors, context }) => {
-  await page.goto("/files/files/");
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
-
-  const shareHash = await page.evaluate(() => localStorage.getItem('shareHash'));
-  if (shareHash == "") {
-    throw new Error("Share hash not found in localStorage");
-  }
-
-  await page.goto("/files/share/" + shareHash + "/testdata/");
-  await expect(page).toHaveTitle("Graham's Filebrowser - Share - testdata");
-  await page.locator('a[aria-label="gray-sample.jpg"]').click({ button: "right" });
-  await page.locator('button[aria-label="Select multiple"]').waitFor({ state: 'visible' });
-  await page.locator('button[aria-label="Select multiple"]').click();
-
-  const popup = page.locator('#popup-notification-content');
-  await popup.waitFor({ state: 'visible' });
-  await expect(popup).toHaveText("Multiple Selection Enabled");
-
-  await page.locator('a[aria-label="20130612_142406.jpg"]').click();
-  await page.locator('a[aria-label="IMG_2578.JPG"]').click();
-  await page.locator('a[aria-label="gray-sample.jpg"]').click({ button: "right" });
-
-  await page.locator('.selected-count-header').waitFor({ state: 'visible' });
-  await expect(page.locator('.selected-count-header')).toHaveText('3 selected');
-
-  await page.locator('button[aria-label="Download"]').waitFor({ state: 'visible' });
-  await page.locator('button[aria-label="Download"]').click();
-  const popup2 = page.locator('#popup-notification-content');
-  await popup2.waitFor({ state: 'visible' });
-  await expect(popup2).toHaveText("Downloading...");
-  checkForErrors();
-});
